Save resultats !

+++ Model/save_model.cpp
@@ -80,12 +80,26 @@ void save_model::saveJsonFile(std::vector<calque> calques, const std::vector<Res
 
 
     for(Resultat i : resultats){
+        //Pertinence
         Json::Value resultValue;
         resultValue["pertinence"] = i.getPertinence();
 
+        //Top left Point
+        Json::Value arrayTopPoint;
+        arrayTopPoint.append(i.getBottomRightPoint().x());
+        arrayTopPoint.append(i.getBottomRightPoint().y());
+        resultValue["topLeftPoint"] = arrayTopPoint;
+
+        //Bottom right Point
+        Json::Value arrayBotPoint;
+        arrayBotPoint.append(i.getTopLeftPoint().x());
+        arrayBotPoint.append(i.getTopLeftPoint().y());
+        resultValue["bottomRightPoint"] = arrayBotPoint;
+
         //Array
         Json::Value arrayV;
-        for (int element: i.getResults()) {
+        //int index = 0;
+        for (auto element: i.getResults()) {
             arrayV.append(element);
         }
         resultValue["resultats"] = arrayV;
@@ -104,12 +118,11 @@ void save_model::saveJsonFile(std::vector<calque> calques, const std::vector<Res
 }
 
 void save_model::save_as(std::string path, std::string fileName, std::vector<calque> _calques,std::vector<Resultat> resultats){
-    std::cout << "function save_as " << std::endl;
+    //std::cout << "function save_as " << std::endl;
 
     auto first = fileName.find(".");
     std::string f = fileName.substr(0, first);
     filename = getFileName(f,true,separator);
-
 
     savePath = path;
 
@@ -117,18 +130,29 @@ void save_model::save_as(std::string path, std::string fileName, std::vector<cal
     saveCalquesPath += separator;
     saveCalquesPath += std::string("Calques");
 
-    if(std::filesystem::exists(saveCalquesPath.c_str())){
-        std::cout << saveCalquesPath << " Found" << std::endl;
+    if(!std::filesystem::is_empty(std::filesystem::path(savePath))){
 
-        std::filesystem::remove_all(std::filesystem::path(saveCalquesPath));
+        if(std::filesystem::exists(saveCalquesPath.c_str())){
+            //std::cout << saveCalquesPath << " Found" << std::endl;
 
+            std::filesystem::remove_all(std::filesystem::path(saveCalquesPath));
 
-        std::cout << saveCalquesPath << " Removed" << std::endl;
+            std::filesystem::remove_all(std::filesystem::path(savePath));
+
+            //std::cout << saveCalquesPath << " Removed" << std::endl;
+
+        }else{
+            savePath = path + separator + filename;
+
+            saveCalquesPath = savePath;
+            saveCalquesPath += separator;
+            saveCalquesPath += std::string("Calques");
+        }
     }
 
 
-    std::filesystem::create_directory(saveCalquesPath.c_str());
-//    fs::remove_all(fs::path(saveCalquesPath));
+
+    std::filesystem::create_directories(saveCalquesPath.c_str());
 
     saveTiff(fileName);
